refactor: getToday 유틸 모듈로 이동
